Supply `float32` to `cv2.remap` in `dewarp` (#28)
* Fix error: (-215:Assertion failed) bug where image coords are of wrong type

* refactor: chain

---------

+++ src/page_dewarp/dewarp.py
@@ -58,13 +58,14 @@ class RemappedImage:
             image_x_coords,
             (width, height),
             interpolation=INTER_CUBIC,
-        )
+        ).astype(np.float32)
         image_y_coords = resize(
             image_y_coords,
             (width, height),
             interpolation=INTER_CUBIC,
-        )
+        ).astype(np.float32)
         img_gray = cvtColor(img, COLOR_RGB2GRAY)
+        # Ensure image_x_coords and image_y_coords are of the correct type
         remapped = remap(
             img_gray,
             image_x_coords,
